requirements.txt
flask
flask-login
python-dotenv
Werkzeug
paho-mqtt